Add menuNavListeners as a module
